Fix with one indirection
Apparently something breaks when trying to supply the asm in nasm as
stdin via /proc/self/fd/0. Hence, we do a single indirection over a file
which somehow makes it work.

Also don't forget to set to 64-bit mode (which is certainly wrong to
hardcode in the end) but works on my machine right now.